lazy load SafeMarkdown component
